requirements.txt
fastapi
uvicorn
pillow
requests
pydantic
azure-identity
azure-keyvault-secrets
azure-storage-blob
pymupdf
nest-asyncio
openai
azure-ai-inference
azure-core
python-multipart
torch == 2.6.0
cohere == 5.13.11
scikit-learn == 1.6.1
python-dotenv == 1.0.1
joblib == 1.3.2

